@@ -1,27 +1,33 @@
-name: netrafit_glasses
-description: "Face Shape Detection & Virtual Glasses Try-On"
+name: face_frame_tryon_frontend
+description: A Flutter app for virtual frame try-on
+
 publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.9.2
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  google_mlkit_face_detection: ^0.13.1
-  google_mlkit_commons: ^0.11.0
-  permission_handler: ^12.0.1
-  camera: ^0.11.3
-  cupertino_icons: ^1.0.8
+  http: ^1.1.0
+  image_picker: ^1.0.4
+  camera: ^0.10.5
+  provider: ^6.1.1
+  shared_preferences: ^2.2.2
+  fluttertoast: ^8.2.2
+  path_provider: ^2.1.1
+  path: ^1.8.3
+  permission_handler: ^11.0.1
+  webview_flutter: ^4.4.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^5.0.0
+  flutter_lints: ^3.0.0
 
 flutter:
   uses-material-design: true
   assets:
-    - assets/
+    - assets/images/
