@@ -1,30 +1,17 @@
-// This is a basic Flutter widget test.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
+// Add this import at the top
+import 'dart:io';
 
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:netrafit/main.dart';
-
-void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
-
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('1'), findsNothing);
-
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
-    await tester.pump();
-
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsNothing);
-    expect(find.text('1'), findsOneWidget);
-  });
+// Add this method to test connection
+Future<void> testServerConnection() async {
+  try {
+    final client = HttpClient();
+    final request = await client.getUrl(Uri.parse('https://violetlike-onward-marley.ngrok-free.dev/'));
+    final response = await request.close();
+    print('✅ Server connection test: ${response.statusCode}');
+    if (response.statusCode == 200) {
+      print('✅ Server is reachable and responding');
+    }
+  } catch (e) {
+    print('❌ Server connection test failed: $e');
+  }
 }
